% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clean_String.R, R/hello.R
\name{Clean_String}
\alias{Clean_String}
\alias{Clean_String}
\title{Una función que limpia una cadena de caracteres, removiendo puntuación y números y tokenizándola}
\usage{
Clean_String(str)

Clean_String(str)
}
\arguments{
\item{str}{Una cadena de caracteres de entrada tal como "This is a cool function!"}

\item{str}{Una cadena de caracteres de entrada tal como "This is a cool function!"}
}
\value{
Un vector que contiene todos los tokens válidos de la cadena de caracteres de entrada original

Un vector que contiene todos los tokens válidos de la cadena de caracteres de entrada original
}
\description{
Una función que limpia una cadena de caracteres, removiendo puntuación y números y tokenizándola

Una función que limpia una cadena de caracteres, removiendo puntuación y números y tokenizándola
}
